Nav got cut off by overflow of the header fix
When you toggled the nav bar the bottom got cut of by the header as the overflow was set to hidden. I created another div and made it just as big as the header. After I added a the background image to the psuedo element and set the position property accordingly. That way the nav doesn't get cut off.